FaceVault/Python/requirements.txt
# Core face recognition and image processing
face_recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3

# Machine learning and clustering
scikit-learn==1.3.0
scipy==1.11.1

# Testing framework
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development and quality tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Utilities
python-dateutil==2.8.2
tqdm==4.65.0